Add flash message and move base layout to layouts folder

+++ src/Controller/PinsController.php
@@ -50,6 +50,8 @@ class PinsController extends AbstractController
             $em->persist($pin);
             $em->flush();
 
+            $this->addFlash('success', 'Pin successfully created!');
+
             return $this->redirectToRoute('app_home');
         }
 
@@ -69,6 +71,8 @@ class PinsController extends AbstractController
 
             $em->flush();
 
+            $this->addFlash('success', 'Pin successfully updated!');
+
             return $this->redirectToRoute('app_home');
         }
 
@@ -81,10 +85,16 @@ class PinsController extends AbstractController
     /**
      * @Route("/pins/{id<[0-9]>}/delete", name="app_pins_delete", methods="DELETE")
      */
-    public function delete(Pin $pin, EntityManagerInterface $em): Response
+    public function delete(Request $request, Pin $pin, EntityManagerInterface $em): Response
     {
-        $em->remove($pin);
-        $em->flush();
+        if ($this->isCsrfTokenValid('pin_deletion_' . $pin->getId(), $request->request->get('csrf_token'))){
+
+            $em->remove($pin);
+            $em->flush();
+
+            $this->addFlash('info', 'Pin successfully deleted!');
+        }
+        
 
         return $this->redirectToRoute('app_home');
     }
